moved profile sections into seperate components to clean things up
